Parse cmap table
