+++ src/fn/index.mjs
@@ -11,10 +11,12 @@ export * from './make-gpos-table.mjs';
 export * from './make-gsub-table.mjs';
 export * from './make-head-table.mjs';
 export * from './make-ltag-table.mjs';
+export * from './make-cmap-table.mjs';
 export * from './make-name-table.mjs';
 export * from './make-os2-table.mjs';
 export * from './make-post-table.mjs';
 export * from './name-table-names.mjs';
+export * from './parse-cmap-table.mjs';
 export * from './parse-fvar-table.mjs';
 export * from './parse-gpos-table.mjs';
 export * from './parse-gsub-table.mjs';
